+++ music/src/main/java/com/music/controller/MusicController.java
@@ -3,8 +3,8 @@ package com.music.controller;
 import com.music.dto.music.MusicReadResponse;
 import com.music.dto.music.MusicUpdateRequest;
 import com.music.dto.music.MusicUpdateResponse;
-import com.music.dto.upload.MusicUploadResponse;
-import com.music.dto.upload.MusicUploadRequest;
+import com.music.dto.music.MusicUploadResponse;
+import com.music.dto.music.MusicUploadRequest;
 import com.music.service.music.MusicReadService;
 import com.music.service.music.MusicUpdateService;
 import com.music.service.music.MusicUploadService;
@@ -15,7 +15,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.bind.annotation.RestController;
