+++ music/src/main/java/com/music/dto/music/MusicUploadResponse.java
@@ -1,10 +1,10 @@
-package com.music.dto.upload;
+package com.music.dto.music;
 
+import com.music.dto.upload.FileKeys;
 import com.music.eneity.Music;
 import com.music.eneity.constants.Genre;
 import com.music.eneity.constants.ReleaseStatus;
 import java.time.LocalDate;
-import javax.persistence.Column;
 import lombok.Builder;
 import lombok.Getter;
